Change starter weapons
In the end, after much playtesting, I found that it wasn't very enjoyable to start with single-action revolvers. It just wasn't appropriate for the enemy numbers. Instead, I think that this is a good place to put handguns that I found myself immediately selling and never using: The pocket pistols, the LCRs, the Union French, and so on.