Reduce sosig health to 50 from 100
In line with changing the starter weapons to pocket pistols, I am reducing sosig health by half. The intention of this character was always to have one-hit kills with any gun, with the challenge being to shoot for gaps in armour quickly and accurately. I think that 50 health should be a good balance between the starter guns being useful (1-2 shots to kill) while still incentivising the player to switch for a more powerful gun that will give them that guaranteed one-hit kill.