Loadout sosigs with handguns/machine pistols only
The loadout pools of each sosig type used to be 2 pistols, a shotgun, and a full-auto weapon (scaling in power and capacity by level), and they could spawn with any one of these. I've changed it so that they can only spawn with similar weapons to you, so that they don't have an inherent advantage over you if you try to play this character on larger maps.