Change starting pool to revolvers from pocket pistols
It was a backwards to have magazine-fed pistols (even the weakest and worst ones) as starter weapons, while the rest of the Lv0 pool was nearly all revolvers. It led to me skipping the revolvers entirely and going into Hold 1 with the starter weapon every time.

I've placed the worst revolvers in this starting pool instead, ones that have slow reloads (Dardick), small capacities (Bull 454), or only one ammo type (OTS-38). These were the pistols I often found myself skipping in the game itself, so hopefully what is left behind is stuff that's more appreciated.